% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callPeaks}
\alias{callPeaks}
\title{call peaks}
\usage{
callPeaks(
  proj,
  groupBy = "Clusters_Combined",
  pathToMacs2 = findMacs2(),
  reducedDims = "LSI_Combined",
  verbose = TRUE
)
}
\arguments{
\item{proj}{An ArchRProject object. The output of \link{processArchR}.}

\item{groupBy}{The groups used to call peaks by function \link[ArchR:addReproduciblePeakSet]{addReproduciblePeakSet}.}

\item{pathToMacs2}{The macs2 path.}

\item{reducedDims}{The reductions used to link the gene and peaks.}

\item{verbose}{Print message or not.}
}
\value{
An ArchRProject object.
}
\description{
call peaks
}
\examples{
# example code

}
