% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContactMatrix.R
\name{createContactMatrix}
\alias{createContactMatrix}
\title{create contact matrix npy file for scDeepLUCIA}
\usage{
createContactMatrix(
  proj,
  pairs_files,
  outdir = "scDeepLUCIA_inputs/sliced_con_array",
  groupBy = "Sample",
  chroms,
  resolution = 5000,
  ...,
  verbose = TRUE,
  use_conda = FALSE
)
}
\arguments{
\item{proj}{An ArchRProject object.}

\item{pairs_files}{File paths for pairs.
The input pairs file must have columnes barcode, chr1, pos1, chr2, pos2}

\item{outdir}{The output folder for contact npy files.}

\item{groupBy}{The column names of cell column data used for split the cells}

\item{chroms}{The chromosome should be worked on.}

\item{resolution}{The resolution for the genomic feature npy file.}

\item{...}{The parameters not used.
See https://github.com/ma-compbio/Higashi/wiki/Configure-the-parameters-for-Fast-Higashi.}

\item{verbose}{Print message or not.}

\item{use_conda}{Use conda to create the python environment or not. If a path
is given, will use that conda env.}
}
\value{
nothing
}
\description{
create contact matrix npy file for scDeepLUCIA
}
\examples{
# example code

}
