% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOnehotGenomicNpy.R
\name{createOnehotGenomicNpy}
\alias{createOnehotGenomicNpy}
\title{create One-hot encoded genomic feature npy file.}
\usage{
createOnehotGenomicNpy(
  proj,
  outdir = "scDeepLUCIA_inputs/sliced_epi_array",
  chroms,
  resolution = 5000
)
}
\arguments{
\item{proj}{The ArchR project}

\item{outdir}{The output directory}

\item{chroms}{The chromosome should be worked on.}

\item{resolution}{The resolution for the genomic feature npy file.}
}
\value{
nothing
}
\description{
create One-hot encoded genomic feature npy file.
}
\examples{
# example code

}
