% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareHigashiFiles.R
\name{prepareHigashiFiles}
\alias{prepareHigashiFiles}
\title{create files required for FastHigashi}
\usage{
prepareHigashiFiles(
  proj,
  pairs_files,
  outputDirectory = "per_cell",
  labelInfoColumns = c("Sample", "Clusters_Combined"),
  plot_label = "Clusters_Combined",
  ...,
  verbose = TRUE
)
}
\arguments{
\item{proj}{An ArchRProject object.}

\item{pairs_files}{File paths for pairs.
The input pairs file must have columnes barcode, chr1, pos1, chr2, pos2}

\item{outputDirectory}{The output folder for filtered pairs and config files
of Higashi.}

\item{labelInfoColumns}{The column names of cell column data used for
label_info.pickle}

\item{plot_label, ...}{The parameters for config JSON.
See https://github.com/ma-compbio/Higashi/wiki/Configure-the-parameters-for-Fast-Higashi.}

\item{verbose}{Print message or not.}
}
\value{
nothing
}
\description{
create files required for FastHigashi
}
\examples{
# example code

}
