% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_RNA_barcode.R
\name{replace_RNA_barcode}
\alias{replace_RNA_barcode}
\title{Replace RNA barcodes to matched ATAC barcodes}
\usage{
replace_RNA_barcode(
  rna,
  bmap,
  barcodesFun = function(x) {
     do.call(rbind, strsplit(colnames(x), "_"))[, 4]
 }
)
}
\arguments{
\item{rna}{A list of Seurat Object.}

\item{bmap}{A named character vector. The names of the vector are
RNA barcode and the elements are DNA barcode.}

\item{barcodesFun}{How to extract the RNA barcode.
The default function is used to extract RNA barcodes for the original manuscript. The barcodes is the cell names in the format of "xw329_day0_120_TACAGCAAACTCAACGTA". If your cell names is in the format of "TACAGCAAACTCAACGTA-1", you may want to try:
barcodesFun = function(x)\{do.call(rbind, strsplit(colnames(x), '-)[1])\}}
}
\value{
A list of Seurat object with metadata DNAbarcode as replaced barcode
}
\description{
Replace RNA barcodes to matched ATAC barcodes
}
\examples{
# example code

}
