% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastHigashi.R
\name{runFastHigashi}
\alias{runFastHigashi}
\title{Run FastHigashi}
\usage{
runFastHigashi(inputDirectory = "per_cell", use_conda = FALSE)
}
\arguments{
\item{inputDirectory}{The file path for Higashi}

\item{use_conda}{Use conda to create the python enviroment or not}
}
\value{
nothing.
}
\description{
Run FastHigashi
}
\examples{
# example code

}
