% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write10X_h5.R
\name{Write10X_h5}
\alias{Write10X_h5}
\title{Write 10X hdf5 file}
\usage{
Write10X_h5(
  seu,
  filename,
  layer = "counts",
  assay = "RNA",
  barcodes = "DNAbarcode"
)
}
\arguments{
\item{seu}{The Seurat object after barcode remapping by \link{replace_RNA_barcode}.}

\item{filename}{Path to h5 file}

\item{layer, assay}{The parameter for \link[SeuratObject:AssayData]{GetAssayData}.}

\item{barcodes}{The colname of barcode in metadata. If it is a Function, such as colnames, The function will be used to extract the  barcode. Otherwise, the column in metadata will be used as barcode.}
}
\value{
nothing
}
\description{
Write count matrix to 10X CellRanger hdf5 file.
This file will be used for \code{ArchR::import10xFeatureMatrix} function.
}
\examples{
# example code


}
