% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArchRproj.R
\name{createArchRproj}
\alias{createArchRproj}
\title{import RNA reads and DNA reads as a ArchR project}
\usage{
createArchRproj(
  rnaFiles,
  atacFiles,
  bmap,
  barcodesFun = function(x) {
     do.call(rbind, strsplit(colnames(x), "_"))[, 4]
 },
  outputDirectory = "ArchR_proj",
  annotationGenes = NULL,
  verbose = TRUE
)
}
\arguments{
\item{rnaFiles, atacFiles}{A named characters.
The file path of RNA and DNA count matrix files.}

\item{bmap, barcodesFun}{The parameter required by \link{replace_RNA_barcode}.}

\item{outputDirectory}{The output folder the the merged RNA and DNA
ArchR project}

\item{annotationGenes}{If there is not standard annotations available,
please input the gene annotations in GRanges object with symbols metadata}

\item{verbose}{Print message or not.}
}
\value{
An ArchRProject object.
}
\description{
import RNA reads and DNA reads as a ArchR project
}
