% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processArchR.R
\name{processArchR}
\alias{processArchR}
\title{process the ATAC/RNA merged ArchR project by standard pipeline}
\usage{
processArchR(
  proj,
  resolutions = c(LSI_ATAC = 0.2, LSI_RNA = 0.2, Clusters_ATAC = 0.4, Clusters_RNA = 0.4,
    Clusters_Combined = 0.4),
  minDist = 0.8,
  skip_rmDup = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{proj}{An ArchRProject object. The output of \link{createArchRproj}.}

\item{resolutions}{The resolutions for each steps.}

\item{minDist}{A number that determines how tightly the UMAP is allowed to pack points together. This argument is passed to min_dist in uwot::umap(). For more info on this see https://jlmelville.github.io/uwot/abparams.html.}

\item{verbose}{Print message or not.}
}
\value{
An ArchRProject object.
}
\description{
process the ATAC/RNA merged ArchR project by standard pipeline
}
\examples{
# example code

}
